backend/requirements.txt
# API
fastapi==0.111.0
pydantic==2.7.1
shortuuid

# MongoDB
beanie==1.26.0
motor==3.4.0

# Prometheus
prometheus-fastapi-instrumentator==7.0.0

# Bot
aiogram==3.7.0

# Auth
python-dotenv==1.0.1
PyJWT==2.8.0